feat: add user track activities

+++ pkg/jwt/jwt.go
@@ -7,7 +7,7 @@ import (
 	"time"
 )
 
-func CreateToken(id int64, username, secretKey string) (string, error) {
+func CreateToken(id uint, username, secretKey string) (string, error) {
 	token := jwt.NewWithClaims(jwt.SigningMethodHS256,
 		jwt.MapClaims{
 			"id":       id,
@@ -26,7 +26,7 @@ func CreateToken(id int64, username, secretKey string) (string, error) {
 	return tokenStr, nil
 }
 
-func ValidateToken(tokenStr, secretKey string) (int64, string, error) {
+func ValidateToken(tokenStr, secretKey string) (uint, string, error) {
 	key := []byte(secretKey)
 	claims := jwt.MapClaims{}
 
@@ -40,7 +40,7 @@ func ValidateToken(tokenStr, secretKey string) (int64, string, error) {
 		return 0, "", errors.New("invalid token")
 	}
 
-	return int64(claims["id"].(float64)), claims["username"].(string), nil
+	return uint(claims["id"].(float64)), claims["username"].(string), nil
 }
 
 func ValidateTokenWithoutExpiry(tokenStr, secretKey string) (int64, string, error) {
